Update SoccerSim.java
Professor Johnson, I have been working on this code for part of Wednesday and all of Thursday and a lot of time before that and I still couldn't get it to pass everything. I genuinely tried my best with the time that I was given and with all of the exams that I had and I'm sorry that I didn't make it perfect. I am really sorry but I did all that I could.